Extract Money Forward Session concept to its own class
This will make it easier to track the session cookie,
pass it around, and build requests that need it.
